
\docType{data}
\name{bore_diameter_data}
\alias{bore_diameter_data}
\title{Bore Diameter in Manufacturing Automotive Driver Gears}
\format{A data frame with 20 rows and 5 variables:
\describe{
   \item{pos1}{Diameter measurement at Position 1}
   \item{pos2}{Diameter measurement at Position 2}
   \item{pos3}{Diameter measurement at Position 3}
   \item{pos4}{Diameter measurement at Position 4}
   \item{pos5}{Diameter measurement at Position 5}
}}
\examples{
data(bore_diameter_data)
}
\description{
Woodall, William H. "Controversies and contradictions in statistical process control." Journal of quality technology 32.4 (2000): 341-350.
  
A dataset containing the diameter of the bore in manufacturing automotive driver gears. The variables are as follows:
}
\keyword{datasets}
