% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grinder_data}
\alias{grinder_data}
\title{Thickness measurement of silicon wafer}
\format{
A data frame with 30 rows and 5 variables:
\describe{
  \item{pos1}{Thickness measurement at Position 1 (outer circle)}
  \item{pos2}{Thickness measurement at Position 2 (outer circle)}
  \item{pos3}{Thickness measurement at Position 3 (middle circle)}
  \item{pos4}{Thickness measurement at Position 4 (middle circle)}
  \item{pos5}{Thickness measurement at Position 5 (inner circle)}
}
}
\usage{
grinder_data
}
\description{
A dataset containing the thickness measurements in nm
at different positions on the silicon wafer
}
\references{
Roes, Kit CB, and Ronald JMM Does. "Shewhart-type charts in nonstandard situations." Technometrics 37.1 (1995): 15-24
}
\keyword{datasets}
