% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pistonring_data}
\alias{pistonring_data}
\title{Pistonring data}
\format{
A data frame with 25 rows and 5 variables:
\describe{
  \item{X1}{Observation 1 in subgroups}
  \item{X2}{Observation 2 in subgroups}
  \item{X3}{Observation 3 in subgroups}
  \item{X4}{Observation 4 in subgroups}
  \item{X5}{Observation 5 in subgroups}
 }
}
\usage{
pistonring_data
}
\description{
A dataset containing piston ring data
}
\references{
Montgomery, Douglas C. 2005. Introduction to Statistical Quality Control. John Wiley & Sons.
}
\keyword{datasets}
