\docType{data}
\name{preston_data}
\alias{preston_data}
\title{Prescription fentanyl consumption in Preston county, WV}
\description{
A dataset containing prescription fentanyl consumption in Preston county, WV, measured using MME percapita. This is a subset from Rich et al. <doi: 10.21105/joss.02450>
}
\examples{
data(preston_data)
}