% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{semiconductor_data}
\alias{semiconductor_data}
\title{Semiconductor data}
\format{
A vector with 50 observations:
\describe{
  \item{obs}{the 151st feature}
 }
}
\usage{
semiconductor_data
}
\description{
A dataset cotaining the 151st feature in SECOM dataset
}
\references{
McCann, Michael, and Adrian Johnston. 2008. “SECOM.” UCI Machine Learning Repository. DOI: https://doi.org/10.24432/C54305.
}
\keyword{datasets}
