\docType{data}
\name{snowfall_data}
\alias{snowfall_data}
\title{Snowfall in inches in Minneapolis/St. Paul, MN}
\description{
A dataset containing snowfalls measured in inches in Minneapolis/St. Paul, MN.  This is a subset of that on the website of Minnesota Department of Natural Resources(https://www.dnr.state.mn.us/climate/twin_cities/snowfall.html)
}
\examples{
data(snowfall_data)
}