
\docType{data}
\name{bore_diameter_data}
\alias{bore_diameter_data}
\title{Bore Diameter in Manufacturing Automotive Driver Gears (mm)}
\format{A data frame with 20 rows and 5 variables:
\describe{
   \item{X1}{Diameter measurement at Position 1}
   \item{X2}{Diameter measurement at Position 2}
   \item{X3}{Diameter measurement at Position 3}
   \item{X4}{Diameter measurement at Position 4}
   \item{X5}{Diameter measurement at Position 5}
}}
\examples{
data(bore_diameter_data)
}
\description{
Source:
Wooluru, Yerriswamy, D. R. Swamy, and P. Nagesh. "THE PROCESS CAPABILITY ANALYSIS-A TOOL FOR PROCESS PERFORMANCE MEASURES AND METRICS-A CASE STUDY." International Journal for Quality Research 8.3 (2014).
  
A dataset containing the diameter of the bore in manufacturing automotive driver gears. The variables are as follows:
}
\keyword{datasets}
