% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bore_diameter_data}
\alias{bore_diameter_data}
\title{Bore diameter in manufacturing automotive driver gears}
\format{
A data frame with 20 rows and 5 variables:
\describe{
  \item{X1}{Diameter measurement at Position 1}
  \item{X2}{Diameter measurement at Position 2}
  \item{X3}{Diameter measurement at Position 3}
  \item{X4}{Diameter measurement at Position 4}
  \item{X5}{Diameter measurement at Position 5}
}
}
\usage{
bore_diameter_data
}
\description{
A dataset cotaining bore diameter measurements in mm
}
\references{
Wooluru, Yerriswamy, D. R. Swamy, and P. Nagesh. "THE PROCESS CAPABILITY ANALYSIS-A TOOL FOR PROCESS PERFORMANCE MEASURES AND METRICS-A CASE STUDY." International Journal for Quality Research 8.3 (2014).
}
\keyword{datasets}
