\docType{data}
\name{grinder_data}
\alias{grinder_data}
\title{Thickness measurement of silicon waffer}
\format{A data frame with 30 rows and 5 variables:
\describe{
   \item{pos1}{Thickness measurement at Position 1}
   \item{pos2}{Thickness measurement at Position 2}
   \item{pos3}{Thickness measurement at Position 3}
   \item{pos4}{Thickness measurement at Position 4}
   \item{pos5}{Thickness measurement at Position 5}
}}
\examples{
data(grinder_data)
}
\description{
A dataset containing the thickness in micrometer at different positions in a grinder. The variables are as follows:
}
\keyword{datasets}
