% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snowfall_data}
\alias{snowfall_data}
\title{Seasonal snowfall in inches in Minneapolis/St. Paul, MN}
\format{
A data frame with 82 rows and 4 variables:
\describe{
  \item{Year}{year of the snowfalls}
  \item{jan}{snowfalls in January}
  \item{feb}{snowfalls in February}
  \item{mar}{snowfalls in March}
 }
}
\usage{
snowfall_data
}
\description{
A dataset containing snowfalls measured in inches in Minneapolis/St. Paul, MN.
}
\references{
Mukherjee, P. S. (2016). On phase II monitoring of the probability distributions of univariate continuous processes. Statistical Papers, 57(2), 539-562.
}
\keyword{datasets}
